Switched from calling the login program to calling bash
Was having issues with getting aliases to set with the login program.
It was nice because it inherited a lot of the current user's state
but ended up being too difficult to configure after the fact.

Instead we can start a bash shell and pass in a customized bashrc file
which allows for total customization